appdata: Migrate to metainfo
Metainfo is the preferred name and location for appdata files.
Since flightgear already ships a metainfo file, and those are
preferred over appdata ones, the computed data in the built
flatpak were the one from the upstream's metainfo, not our
appdata with all the screenshots. So now the upstream metainfo
is replaced by the one in this flathub repository, which means
the new screenshots are now shown on both the flathub.org app
page and in GNOME Software et al.