manifest: Drop ffmpeg dependency
I wasn't sure at that time if flite/flightgear really
depended on it, but I checked again and saw no signifiant
mention of ffmpeg in both. I checked with ldd too and
it didn't show that it linked against it anyway.
Since the ffmpeg runtime is quite heavy, it's better not to
depend on it if it's not necessary.

So this commit drops the ffmpeg flatpak runtime dependency
altogether.